data: Docs for encrypted volumes for compute machines
Partially reverts 0c370dd811 (data/aws: Encrypt the AMI used by the
bootstrap and master machines, 2019-02-22, #1296).  This isn't a clean
revert; for example, I left the ability to destroy images which are
tagged as owned by the cluster.  And we're still copy-and-encrypting
for the bootstrap machine and control-plane machines until the AWS
Terraform provider supports requesting encrypted root volumes [1].
But with this commit, we're now documenting the encryption in a way
that covers both the previous AMI-based encryption used for
bootstrap/control-plane and the new root-volume-based encryption used
for the compute machines, because they come down to encrypted root
volumes regardless of their approach.

[1]: https://github.com/terraform-providers/terraform-provider-aws/issues/8624