openstack: unhardcode the cloud name
To allow users to provide custom cloud names in the install
config we need to remove the hardcoded 'openstack' name.

To do so and to keep the compatibility with the current behavior,
we read cloud name from the install config first, and if it's empty,
set it to the value of OS_CLOUD env variable. If both values are not
specified, then cloud name defaults to "openstack".