openstack: Remove the VIPs from the Platform
The `APIVIP`, `DNSVIP` and `IngressVIP` values not supposed to be
user-configurable. They're set by the installer and setting them to
a value other than the expected one will likely result in an error.

Therefore, this removes them from `InstallConfig.Platform` in favour
of free functions that return the expected values.