Install sshuttle in openstack CI image
The node that runs the installer need to access a DNS resolver that has
the A api records for the CI provisioned clusters. This resolver is
currently hosted in MOC and wide open but will likely need to be locked
down. `sshuttle` can capture the DNS queries and forward them to the
CI-DNS VM on MOC, provided it has SSH access.

https://sshuttle.readthedocs.io/en/stable/manpage.html#cmdoption-sshuttle-dns