[gcp] pkg/destroy/gcp: Use email to destroy service accounts
We were not initially using email because infra ID could be larger than
what the credential operator was allocating for it in the email (max 30
chars). However, now we are limiting gcp infra ID's to a much smaller
size that does fit in the email. We are not guaranteed that a
displayName will be set on a service account, but we are guaranteed that
an email will be set.