pkg/destroy/aws/aws: Return ARNs from terminateEC2InstancesByTags
We might want this same sort of thing in deleteEC2NATGatewaysByVPC and
our other functions where the caller doesn't know what resources the
function might be deleting.  But at the moment, Abhinav is only asking
for a preemptive guard for doubled instance investigation [1].

The ARN syntax for instances is [2]:

  arn:aws:ec2:{region}:{account-id}:instance/{instance-id}

and while we have to stretch into the ec2Client and reservation
objects to find the region and account ID, constructing the ARN
ourselves isn't that bad.

Without this guard, any instances that terminateEC2InstancesByTags
terminates but which still show up via the generic tag search will
flow down through Run's deleteARN, through terminateEC2Instance, and
end up in terminateEC2InstanceByInstance where terminated instances
are silently skipped.  So it doesn't save us all that much to return
terminated instance ARNs, just the DescribeInstances call in
terminateEC2Instance.

[1]: https://github.com/openshift/installer/pull/2169#discussion_r312046666
[2]: https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html#arn-syntax-ec2