openstack: have gather bootstrap look for FIP
OpenStack installations may have floating IPs that get used for
bootstrap nodes. This patch updates the bootstrap IP code so that
it looks for the floating IP first. If found it will be used.
Otherwised it can be silently ignored in favor of direct
access via the openstack instances access_ipv4 address.

As OpenShift CI for OpenStack uses floating IPs this should resolve
issues in obtaining bootstrap logs in our CI scripts.