hack: remove compile time RELEASE_IMAGE override
The location of the defaultReleaseImage variable has moved to
pkg/asset/releaseimage, and this doesn't currently work. It also appears
nothing is relying on this feature so we can remove it.  If you want to
override the release image, you can do at run time with
OPENSHIFT_RELEASE_IMAGE_OVERRIDE, which functions the same other than
not being embedded in the produced openshift-install binary.