docs/user/aws/install_upi: Document bring-your-own-DNS
Some users want to provide their own *.apps DNS records instead of
delegating that to the ingress operator [1].  With this commit, we
tell the ingress operator not to worry about managing any hosted
zones, and walk users through how they can create the expected records
[2] themselves.

Removing the zones from the YAML manifest via sed or other POSIX
command was too complicated, so I've given up on that and moved to
Python and PyYAML [3].  There are many possible alternatives, but
PyYAML seemed the most likely to be already installed, it's packaged
for many systems if users want to install it, and the syntax is fairly
readable if users want to accomplish the same task with a different
tool of their choice.  The Python examples are more readable as
multi-line strings than if they were one-liners, and they can still be
copy-pasted into a shell.  Once faq [4] or similar becomes more common
on user systems, we can replace this with:

  $ DATA="$(faq '.compute[0].replicas=0' install-config.yaml)"
  $ echo "${DATA}" >install-config.yaml

and similar.

For not, I'm not suggesting admins monitor for other DNSRecord objects
[5] and fullful them as they show up.  In case we do decide to have
folks monitor them later, here's a sample:

  $ oc -n openshift-ingress-operator get -o yaml dnsrecord default-wildcard
  apiVersion: ingress.operator.openshift.io/v1
  kind: DNSRecord
  metadata:
    creationTimestamp: "2019-08-22T20:45:00Z"
    finalizers:
    - operator.openshift.io/ingress-dns
    generation: 1
    labels:
      ingresscontroller.operator.openshift.io/owning-ingresscontroller: default
    name: default-wildcard
    namespace: openshift-ingress-operator
    ownerReferences:
    - apiVersion: operator.openshift.io/v1
      blockOwnerDeletion: true
      controller: true
      kind: IngressController
      name: default
      uid: b31db6db-c51d-11e9-8a7a-02ae97362ddc
    resourceVersion: "8847"
    selfLink: /apis/ingress.operator.openshift.io/v1/namespaces/openshift-ingress-operator/dnsrecords/default-wildcard
    uid: b59fbbfa-c51d-11e9-8a7a-02ae97362ddc
  spec:
    dnsName: '*.apps.wking.devcluster.openshift.com.'
    recordType: CNAME
    targets:
    - ab37f072ec51d11e98a7a02ae97362dd-240922428.us-west-2.elb.amazonaws.com
  status:
    zones:
    - dnsZone:
        tags:
          Name: wking-nfnsr-int
          kubernetes.io/cluster/wking-nfnsr: owned
    - dnsZone:
        id: Z3URY6TWQ91KVV

The route listing is from a cluster running [6].

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1715635
[2]: https://github.com/openshift/cluster-ingress-operator/blob/9ce86811e6aeea58d3b2b8955591d50cd9311885/pkg/operator/controller/ingress/dns.go#L75-L115
[3]: https://pyyaml.org/
[4]: https://github.com/jzelinskie/faq
[5]: https://github.com/openshift/cluster-ingress-operator/blob/d115a146611871c47f44f239346ecea0015993d9/pkg/api/v1/types.go#L18-L25
[6]: https://mirror.openshift.com/pub/openshift-v4/clients/ocp-dev-preview/4.2.0-0.nightly-2019-08-25-233755/