baremetal: Use podman inspect to check ironic service status
Some people are hitting issues where the containers appear running in
podman ps output, but are in fact unresponsive and podman exec/inspect
CLI options fail.

This may be a libpod bug (looking for related issues), but as a workaround
we can check the inspect status, which should mean we can detect zombie
containers and restart the ironic.service which appears to solve the
issue.

Related: https://github.com/openshift-metal3/dev-scripts/issues/753