OpenStack: get rid of swift temp urls
Since not all the configurations of OSP support Swift temp urls, we
have to find another way that will work on all supported clouds.

To create a unified solution we can use the ACL mechanism to allow
public access to the file, but deny access to the list of container files.
From a technical point of view, it looks like this: ".r:*", but no ".rlistings".
This allows anybody to download the ignition file. However, to do this, the
exact name of the object must be known since users cannot list the objects in
the container.

The last step is to randomize the file name so that it cannot be matched - now it
is hardcoded to "bootstrap.ign"

The lifetime of such a file will be short, about 10 minutes necessary for
bootstrapping. Further this file will be automatically removed by Terraform at
the moment of bootstrap machine destruction.

This way we can achieve the same effect as temp url, which will work on all types
of storage.