Bug 1746748: OpenStack: idempotent cluster deletion
Now we delete OpenStack resources in parallel using goroutines.
It can lead to the fact that a resource can already been deleted
when another goroutine sends a request.

Instead of returning an error in this case and quitting the installer
we should tolerate this type of errors and continue to remove resources.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1746748