docs/user/customization: Add a section on imageContentSources
Leaning on the design doc in
docs/dev/alternative_release_image_sources.md, which I'm not linking
because it's more focused on design/internals than I think users need
to be.  Also leaning on [1] for the 'oc adm release mirror'
invocation.  Tested with:

  $ oc adm release mirror --from=quay.io/openshift-release-dev/ocp-release-nightly:4.2.0-0.nightly-2019-09-11-114314 --to=registry.svc.ci.openshift.org/wking/openshift4 --to-release-image=registry.svc.ci.openshift.org/wking/openshift4:4.2.0-0.nightly-2019-09-11-114314

using the oc extracted from that same nightly.

The XXXXXX fuzzing is at Ahbinav's request [2] to block blind
copy/pasting.

[1]: https://github.com/christianh814/blogs/tree/278ac1cf8268511c1d6cad9aaaace62e681d573e/docs/openshift-4.2-disconnected#mirroring-images
[2]: https://github.com/openshift/installer/pull/2356/files#r324377882