baremetal: Add etcd SRV record lookup
On baremetal we don't know the names of the etcd nodes ahead of time
because they are assigned by the customer's DHCP. As a result, we
can't populate the list of etcd cluster members before bootstrap and
need to do the lookup during bootstrap.

This change moves the MCO bootstrap earlier in the process because
that is what configures the nodes to start advertising their DNS
names. It also adds a baremetal-specific call to baremetal-runtimecfg
to do the lookup of the etcd cluster names.

Note that currently we get around this by requiring specific names
to be configured in the customer DHCP, which is rather rigid and
requires more configuration on their side than is desirable. This
change improves the baremetal install experience by removing that
requirement.