pkg/gather/ssh: Add logging for agent generation
Log the choice to use SSH_AUTH_SOCK, since that will clarify cases
where we fail to gather logs because a pre-existing agent lacks the
required SSH key.

Adjust the loading to return a map, so we can keep the source of the
private key around to log in newAgent.  Would be nice to log
fingerprints too, but I couldn't figure out how to convert the
private-key interface{} to a fingerprint.