gcp upi: allow nodePort between masters and workers
Before this change, the gcp upi firewall rules limited access to the
nodePort ports from worker to worker and master to master. Access
between worker and master was denied. However, because gcp upi produces
masters that have the 'worker' role, the nodePort services could run on
masters and need to be accessed from pods on other workers. Or pods on
masters might need access to nodePort services on workers.

This change modifies the gcp upi firewall rules to allow the nodePort
services across the entire cluster.