cmd/openshift-install/destroy: Remove terraform.tfstate in 'destroy cluster'
So 'destroy cluster' gets you all the way back to a clean slate, even
if the cluster in question died during infrastructure provisioning.

This will obviously not clean up the original asset directory in
workflows where the user copies their metadata.json over into a new
directory and runs 'destroy cluster' in the new directory.  But since
we have existing asset-state removal code that also behaves that way,
I don't think it's a big deal.

cmd/openshift-install/destroy.go is a convenient place to put this
now, when all of our providers are Terraform-based.  If, in the
future, we move some providers off of Terraform (or add new,
non-Terraform providers), we can push this down into the
platform-specific destroyers.  We could also leave it here, because
terraform.tfstate is unlikely to exist in the asset directory for
non-Terraform providers and be something that the user wants to keep
around, so the risk of false-positive removal is low.