upi/aws/cloudformation/02_cluster_infra: "the External" -> "the external", etc.
Copy-edits for some output descriptions.  Internal and external are
not proper nouns or otherwise in need of special capitalization.  Also
insert some missing articles.  And drop the trailing "created" (short
for "created by this template"?), because all of these outputs are
related to resources generated by this template, otherwise we wouldn't
need to output them.