cmd/openshift-install/gather: Gather ClusterOperator errors
We currently show ClusterVersion messages when we time out, based on
the expectation that the cluster-version operator (CVO) would bubble
up and summarize any operator-level errors.  But while the CVO does
expose the fact that there *was* an error, it sometimes simplifies to
the point where users can no longer determine the underlying condition
[1], while the ClusterOperator conditions are more useful [2].  With
this commit, we'll gather and log surprising ClusterOperator
conditions (Available=false, Degraded=true, etc.) to give the user a
bit more information to dig into before they drop into
'openshift-install gather bootstrap ...' or 'oc adm must-gather ...'.

I personally prefer structured logging for this:

  logger.WithFields(logrus.Fields{
    "condition": condition.Type,
    "status":    condition.Status,
    "reason":    condition.Reason,
  }).Info(condition.Message)

But Abhinav felt that logging a string was more readable [3].  I'd
also rather log Upgradeable=False, because that suggests something
concerning about the reporting operator, but Abhinav felt it would be
distracting [4], and it's certainly not as important as Degraded=True.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1754939#c10
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1754939#c16
[3]: https://github.com/openshift/installer/pull/2450#discussion_r330867075
[4]: https://github.com/openshift/installer/pull/2450#discussion_r330867229