pkg/asset/cluster/aws: Tag shared subnets
So the kubelet cloud-provider can place load balancers appropriately.

This also helps make it clear to other consumers that we are using
these resources for the cluster ("hmm, I wonder if it would be a bad
idea to delete this subnet?  Ah, yeah, I don't want to break cluster
example-123").  But because AWS only allows 50 tags per resource [1],
we're not tagging shared resources like the VPC or Route 53 zones
because those don't need to be tagged to get a working cluster.  Users
are free to add 'shared' tags to resources like that as they see fit.
I also considered using the EC2-specific API [2], but went with the
generic tag API in case we wanted to tag Route 53 zones or other
non-EC2 resources as well in the future.

Tag before entering Terraform to claim the space before we actually
put cluster resources inside the subnets.  That keeps folks from
unknowingly reaping them before we start adding resources, and it
ensures we have space for the new tags.

The 20-tag limit that leads to the loop is from [3].  We're unlikely
to exceed 20 with just the VPC and subnets, but it seemed reasonable
to plan for the future to avoid surprises if we grow this list going
forward.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html#tag-restrictions
[2]: https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateTags.html
[3]: https://docs.aws.amazon.com/sdk-for-go/api/service/resourcegroupstaggingapi/#TagResourcesInput