Revert "pkg/destroy: data/aws: delete undiscoverable AWS objects"
This reverts commit 6ae7598063863c2f318b137e1b2354aef2514fc1, #1268.

That was a workaround to recover from openshift-dev clusters where an
in-house pruner is removing instances but not their associated
instance profiles.  Folks using the installer's destroy code won't
need it, and while the risk of accidental name collision is low, I
don't think it's worth taking that risk.  With this commit, folks
using external reapers are responsible for ensuring that they reap
instance profiles when they reap instances, and we get deletion logic
that is easier to explain to folks mixing multiple clusters in the
same account.