azure: set the os disk type based on machine objects
Like most of the other platforms, the disk type for master machines is present in the machine object, and the terraform should
use that to create control plane machines.

This allows users to use multi-stage installs to modify the control-plane root volume disk type.

This change explicitly skips the addition of making it easy for users to specify this setting in install-config for future work.