data/aws/vpc: Add an elasticloadbalancing endpoint
To avoid [1]:

  The DNS provider failed to ensure the record: failed to get hosted
  zone for load balancer target
  "afdffb7c0550244d39ac35bdfd21fff0-1591348991.us-east-1.elb.amazonaws.com":
  failed to describe load balancers: RequestError: send request failed
  caused by: Post
  https://elasticloadbalancing.us-east-1.amazonaws.com/: dial tcp
  54.239.29.168:443: i/o timeout

when running the DNS provider in a blackhole where it can't get out to
the wider internet.  With the LB endpoint exposed inside the VPC, the
operator will be able to reach the cloud provider without needing to
go through the proxy.  I expect we will want the ingress operator to
respect proxy settings anyway, but at least with this change we won't
*need* that for blackholed private subnets.

docs/user/aws/images/install_upi.dia needs to be updated to reflect
the new endpoint, but I couldn't figure that out on short notice and
it seems safe to punt.  There's also an issue with tagging [1]:

  The DNS provider failed to ensure the record: failed to find hosted
  zone for record: failed to get tagged resources: RequestError: send
  request failed caused by: Post
  https://tagging.us-east-1.amazonaws.com/: dial tcp 52.94.233.76:443:
  i/o timeout

but apparently no VPC endpoints for tagging.  But one step at a time
;).

Tagging these VPC endpoints themselves would be nice too [2], but the
AWS Terraform provider grew support for that in 2.16.0 [3], and we're
still at 2.10.0 per pkg/terraform/exec/plugins/Gopkg.lock.

[1]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_release/5308/rehearse-5308-pull-ci-openshift-installer-master-e2e-aws-proxy/10/artifacts/e2e-aws-proxy/must-gather/namespaces/openshift-ingress-operator/ingress.operator.openshift.io/dnsrecords/default-wildcard.yaml
[2]: https://aws.amazon.com/about-aws/whats-new/2019/05/amazon-vpc-endpoints-now-support-tagging-for-gateway-endpoints-interface-endpoints-and-endpoint-services/
[3]: https://github.com/terraform-providers/terraform-provider-aws/commit/864285b1bb6c014e2352c17d714969e9b9624712