pkg/asset/installconfig/aws/metadata: Store AWS metadata
This allows us to hold the AWS session (and in the future, other
metadata) on the InstallConfig object, where we can load it once in
InstallConfig.Generate or InstallConfig.Load, instead of having to
call GetSession or other loaders in all consumers.  At the moment, I'm
just stubbing in the framework; I'll port our other GetSession
consumers in future commits.

I'm using private properties with getters so we don't have to pay the
price of initializing properties we don't need to use.  Consumers can
call the getter, which will pull fresh data into the caching property
the first time, and return the previously cached data on subsequent
calls.