Azure: Create clusters with no public endpoints
This gives the install the ability to install an OpenShift cluster on Azure as
internal/private, which is only accessable from an internal network and not
visible on the Internet.

Because of limitations in Azure, VM's without a public IP address behind a load
balancer without a public IP address are unable to reach the internet. To get
around this limitation, a dummy load balancer with a public IP address is setup
with a rule to a dummy service.

https://jira.coreos.com/browse/CORS-1227