data/bootstrap/files/usr/local/bin/installer-*gather: Gather failed systemd units
To help in cases where we die before getting far enough to launch a
kubelet (or we have an issue with some other systemd-managed service).
In cases where there are no failing units, this will only add a small
file and an empty directory to the gathered tarball.

The POSIX locale [1] gets us asterix prefixes in the list-units
output, which are easier to match with sed than the default U+25CF
(BLACK CIRCLE) it prints with LANG=en_US.UTF-8.  I'm not actually sure
what locale we usually run on our control-plane machines, but it seems
safest to not leave that up to the machines in this case.
Unfortunately I haven't found a way to have list-units spit out a more
structured output format.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap07.html#tag_07_02