Add OpenStack UPI Documentation
This documents the process of deploying OpenShift on OpenStack using
User Provided Infrastructure. This commit lists all the necessary
steps manually, using the standard `openstack` command to create the
resources.

We intend to provide automation in subsequent commits.

In the meantime, this should provide enough information to do a UPI
deployment either manually or script it however they like.