Use an internal ingress by default when the cluster is marked internal
When the cluster is internal according to the install config's publishing field,
render an internal default ingresscontroller so that ingress remains consistent
with the intent for the entire cluster to be "internal".

This commit is part of the "User-defined Default Ingress Controller" enhancement
proposal[1] and implements the final user-facing component of the feature.

[1] https://github.com/openshift/enhancements/blob/master/enhancements/user-defined-default-ingress-controller.md