[fcos] bootstrap: Copy MCD, disable zincati, enable GCP-routes and pivot
In OKD-on-FCOS, the bootstrap host runs FCOS and does not have
the MCD binary included by default. This change pulls the binary
from the MCO payload image.

Zincati, the default update agent of Fedora CoreOS is disabled
because it is not used in OKD.

On GCP, the gcp-routes.service is enabled to manage routes in
concert with the gcp-routes-controller container.

It is then pivoted into the machine-os-content payload on the
bootstrap host before proceeding with the bootstrapping of the
cluster in order to get a base OS with all of OKD's dependencies.