Filter by tag servers that should be destroyed
Now to delete servers we fetch all existing servers from Nova and
iterate through them, looking for the right ones based on the tag
information.
This approach has drawbacks, because in the case of a large number
of servers, such a request can take a long time.

This commit filters server by tags and thus only downloads the ones
we need.