cmd/openshift-install/gather: --bootstrap and --master override Terraform state
When command line options are set, don't clobber them from the
Terraform state.  In some cases, this means we won't have to hit
Terraform state at all, which gives us a safety valve if we find a
Terraform state file but for some reason (installer version mismatch?)
fail to parse it.

Also validate the values we get to ensure that they are
syntatically-valid IP addresses or domain names.  I'm not including
the validate.IP error, because it doesn't say anything beyond "%q is
not a valid IP", and we're already saying that.