pkg/validate: '%s' -> %q for IP quoting
Save a few characters and get safer quoting if the user gives us a
candidate IP address that contains a single quote or some such.

The error messages I'm touching in the unit tests are fairly
redundant, but I didn't drop their inclusion of the validate.IP error
in case we make the error messages more specific later.