vendor: update terraform-provider-ironic to v0.1.8
Recent versions of Ironic support sending the config drive data as a
JSON blob. Previously we built the ISO on the host installing the
cluster, requiring a dependency on mkisofs being installed. This release
of the terraform provider will just send the user data and related
information directly to Ironic and let it build the ISO, if the Ironic
version is new enough.