data/aws: Shift etcd DNS records from route53 into bootstrap
These should be just for etcd cluster formation [1].  Moving them to
the bootstrap module means they'll be deleted after bootstrapping
completes (by which point we have successfully formed the etcd
cluster).  And the removal simplifies the post-install cluster
environment and makes it clear that there are no in-cluster components
which are leaning on these DNS records.

[1]: https://github.com/etcd-io/etcd/blob/master/Documentation/op-guide/clustering.md#dns-discovery