Add a `fips` flag to install-config.yaml
Part of: https://github.com/openshift/enhancements/pull/15

We added FIPS to the MCO a while ago:
https://github.com/openshift/machine-config-operator/pull/889

However, during some discussion it became clear that the main
use case for FIPS is "day 1" - it doesn't make sense to turn it
on "day 2" because the standard requires that e.g. long-term key
material was created with FIPS enabled.

Further, it's unlikely that admins will want to turn it *off*
if they ever had it on.

This is a good candidate for an install config.