manifests/gcp: add infra-id as prefix for external IGs
To allow re-using the instance groups that control-plane machines already belong to for backend of internal LBs, the upstream PR [1]
requires we set the external-instance-groups-prefix to filter the instance groups for the clusters.

The upstream PR [1] is in flight, but origin backport [2] might actually be merged for now for the time we work with upstream to support the use-case.

[1]: https://github.com/kubernetes/kubernetes/pull/84466
[2]: https://github.com/openshift/origin/pull/24036